carve out custom lnd stuff from custom_grid_widget
